tasks/extract-data/requirements.txt
functions-framework>=3.10.1
geopandas>=1.0.1
google-cloud-storage>=3.9.0
google-cloud-bigquery>=3.41.0
pandas>=2.3.3
pip>=26.0.1
pyarrow>=21.0.0
requests>=2.32.5
